move http function to http.js
