+++ packages/salesforce/src/index.js
@@ -6,3 +6,4 @@ export default Adaptor;
 export * from './Adaptor';
 
 export * as util from './util';
+export * as http from './http';
